aider: Modified SQL query to fetch writer's data based on the `WriterID` in the `edit_writer` route.
# Context:
USER: http://127.0.0.1:5000/edit_writer/1 is not populating with record data
ASSISTANT: It seems like the issue might be with the retrieval of the writer's data in the `edit_writer` route in the `app.py` file. Let's modify the SQL query to fetch the writer's data based on the `WriterID`.

Here is the change you need to make:

```python
app.py
<<<<<<< ORIGINAL
    cursor.execute(''' SELECT * FROM writers WHERE WriterID = %s ''', (WriterID,))
=======
    cursor.execute(''' SELECT * FROM writers WHERE WriterID = %s ''', [WriterID])
>>>>>>> UPDATED
```

In the original code, the `WriterID` is passed as a single value to the `execute` method. In the updated code, we pass the `WriterID` inside a list. This is because the `execute` method expects the parameters to be in a sequence (like a list or tuple), even if there is only one parameter.