aider: Changed the way writer's data is fetched from the database in app.py and modified the way writer's data is accessed in edit_writer.html.
# Context:
USER: the query is returning data but  it is not being populated in the fields in edit_writer.html. the variables migh tneed to be updated to indexes
ASSISTANT: To fix this issue, we need to modify the `app.py` file and the `templates\edit_writer.html` file.

In `app.py`, we need to change the way we fetch the writer's data from the database. Instead of fetching it as a tuple, we should fetch it as a dictionary. This will allow us to access the data using the column names in the `edit_writer.html` file.

In `templates\edit_writer.html`, we need to change the way we access the writer's data. Instead of using dot notation (e.g., `writer.Task`), we should use bracket notation (e.g., `writer['Task']`). This is because we are now fetching the data as a dictionary.

Here are the changes:

```python
app.py
<<<<<<< ORIGINAL
    cursor = mysql.connection.cursor()
    cursor.execute(''' SELECT * FROM writers WHERE WriterID = %s ''', [WriterID])
    writer = cursor.fetchone()
=======
    cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
    cursor.execute(''' SELECT * FROM writers WHERE WriterID = %s ''', [WriterID])
    writer = cursor.fetchone()
>>>>>>> UPDATED
```

```html
templates\edit_writer.html
<<<<<<< ORIGINAL
        <input type="text" id="task" name="task" value="{{ writer.Task }}"><br>
        <input type="text" id="topic" name="topic" value="{{ writer.Topic }}"><br>
        <input type="text" id="style" name="style" value="{{ writer.Style }}"><br>
        <input type="text" id="audience" name="audience" value="{{ writer.Audience }}"><br>
        <input type="text" id="length" name="length" value="{{ writer.Length }}"><br>
        <input type="text" id="format" name="format" value="{{ writer.Format }}"><br>
        <input type="text" id="additional_information" name="additional_information" value="{{ writer.AdditionalInformation }}"><br>
=======
        <input type="text" id="task" name="task" value="{{ writer['Task'] }}"><br>
        <input type="text" id="topic" name="topic" value="{{ writer['Topic'] }}"><br>
        <input type="text" id="style" name="style" value="{{ writer['Style'] }}"><br>
        <input type="text" id="audience" name="audience" value="{{ writer['Audience'] }}"><br>
        <input type="text" id="length" name="length" value="{{ writer['Length'] }}"><br>
        <input type="text" id="format" name="format" value="{{ writer['Format'] }}"><br>
        <input type="text" id="additional_information" name="additional_information" value="{{ writer['AdditionalInformation'] }}"><br>
>>>>>>> UPDATED
```

Please let me know if these changes solve your issue.