aider: Change the form action for the EDIT button on writers.html to a GET method.
# Context:
USER: update the router for edit_writer so that the EDIT button on writers.html routes to edit_writer.html
ASSISTANT: To update the router for `edit_writer` so that the EDIT button on `writers.html` routes to `edit_writer.html`, we need to modify the `writers.html` file.

The form action for the EDIT button needs to be changed from a POST method to a GET method since we are not submitting any data but requesting the `edit_writer` page.

Here is the change:

```html
templates\writers.html
<<<<<<< ORIGINAL
        <form action="{{ url_for('edit_writer', WriterID=writer.WriterID) }}" method="post" style="display: inline;">
            <input type="hidden" name="_method" value="EDIT">
            <button type="submit" class="edit-button">Edit</button>
        </form>
=======
        <form action="{{ url_for('edit_writer', WriterID=writer.WriterID) }}" method="get" style="display: inline;">
            <button type="submit" class="edit-button">Edit</button>
        </form>
>>>>>>> UPDATED
```

This change will make the EDIT button route to the `edit_writer` page when clicked.