aider: Fix issue with constructing the `prompt` variable by checking if all the required keys are present in the dictionaries.
# Context:
USER: the data describing the writer is not being transferred to the 'Completed Prompt' box
ASSISTANT: The issue seems to be with the way the `prompt` variable is being constructed. The `prompt` variable is constructed using the `project` and `selected_writer` dictionaries. If any of the keys used to construct the `prompt` variable are not present in these dictionaries, the `prompt` variable will not be constructed correctly.

To fix this, we need to ensure that all the keys used to construct the `prompt` variable are present in the `project` and `selected_writer` dictionaries. We can do this by adding checks before the line where the `prompt` variable is constructed.

Here's how you can modify the `app.py` file:

```python
app.py
<<<<<<< ORIGINAL
if 'Task' in project and 'Style' in selected_writer:
    prompt = f"Task: {project['Task']}\nTopic: {project['Topic']}\nStyle: {selected_writer['Style']}\nTone: {selected_writer['Tone']}\nAudience: {selected_writer['Audience']}\nLength: {selected_writer['Length']}\nFormat: {selected_writer['Format']}\nAdditional Information: {selected_writer['AdditionalInformation']}"
else:
    prompt = "Incomplete data. Please ensure all fields are filled."
=======
if all(key in project for key in ['Task', 'Topic']) and all(key in selected_writer for key in ['Style', 'Tone', 'Audience', 'Length', 'Format', 'AdditionalInformation']):
    prompt = f"Task: {project['Task']}\nTopic: {project['Topic']}\nStyle: {selected_writer['Style']}\nTone: {selected_writer['Tone']}\nAudience: {selected_writer['Audience']}\nLength: {selected_writer['Length']}\nFormat: {selected_writer['Format']}\nAdditional Information: {selected_writer['AdditionalInformation']}"
else:
    prompt = "Incomplete data. Please ensure all fields are filled."
>>>>>>> UPDATED
```

This change will check if all the keys used to construct the `prompt` variable are present in the `project` and `selected_writer` dictionaries before trying to construct the `prompt` variable. If any of the keys are not present, it will set the `prompt` to a default message.