aider: Added "Tone" field to writers.html table headers and table data.
# Context:
USER: add field to writers.html
ASSISTANT: To add the "Tone" field to the writers table in the `writers.html` file, we need to modify the table headers and the table data.

Here's the change:

```html
templates\writers.html
<<<<<<< ORIGINAL
            <th style="padding: 15px; text-align: left;">Style</th>
            <th style="padding: 15px; text-align: left;">Audience</th>
=======
            <th style="padding: 15px; text-align: left;">Style</th>
            <th style="padding: 15px; text-align: left;">Tone</th>
            <th style="padding: 15px; text-align: left;">Audience</th>
>>>>>>> UPDATED
```

This change adds a new table header for "Tone" after the "Style" header.

```html
templates\writers.html
<<<<<<< ORIGINAL
            <td style="padding: 15px; text-align: left;">{{ writer.Style }}</td>
            <td style="padding: 15px; text-align: left;">{{ writer.Audience }}</td>
=======
            <td style="padding: 15px; text-align: left;">{{ writer.Style }}</td>
            <td style="padding: 15px; text-align: left;">{{ writer.Tone }}</td>
            <td style="padding: 15px; text-align: left;">{{ writer.Audience }}</td>
>>>>>>> UPDATED
```

This change adds a new table data cell for the writer's "Tone" after the "Style" cell. The cell displays the value of the writer's "Tone" field.